server/requirements.txt
Flask
mysql-connector
tabulate
